files_test.go formats_test.go copy_test.go license_test.go helpers.go testdir.go
